Add coverage for alter table concurrent with updates and scans
This adds two threads to the alter table integration test: one which
performs concurrent updates and another which performs concurrent scans.

This triggers various concurrency issues in alter table, so in order
to keep tests green, this also disables the test until we can get it
to pass reliably.

This commit also refactors the WrappedFlush() method which is used in
a few tests into a client-test-util.h module. There is a slight
behavior change here in that the old code used to use ASSERT in some places
and CHECK in others, whereas the new code always uses CHECK (so it's safe
to use from non-main threads)

Change-Id: Ib4c76016272f19bd796cb793a5881e4e7c2574d4