Test for KUDU-462: tablet fails to bootstrap after a column is dropped
This adds a test (but not a fix) for a bug where the tablet fails to
bootstrap after any column is dropped. The issue is that we try to
decode writes against the current schema instead of the schema at the
time of the operation, and thus fail when a column is missing.

Since the test currently fails, it is disabled.

Change-Id: Icc65a0862c7af60df4efe95dc122c9d7b13d8b34