plain_block: pre-reserve space in our buffer
Looking at CPU usage in a cluster test, we see a lot of stacks like:

  Thread 201 (Thread 0x7f7c35b73700 (LWP 10830)):
  #0  0x00000036b2289b8c in memcpy () from /lib64/libc.so.6
  #1  0x00007f7c369648eb in kudu::faststring::GrowArray(unsigned long) () from /var/lib/jenkins/cdep/kudu/build/release/libkudu_util.so
  #2  0x00007f7c378b5dbb in kudu::cfile::PlainBlockBuilder<(kudu::DataType)7>::Add(unsigned char const*, unsigned long) () from /var/lib/jenkins/cdep/kudu/build/release/libcfile.so
  #3  0x00007f7c378a246a in kudu::cfile::CFileWriter::AppendEntries(void const*, unsigned long) () from /var/lib/jenkins/cdep/kudu/build/release/libcfile.so

It appears we weren't pre-sizing the buffer for PlainBlockBuilder, which caused
a lot of allocations/copies in the case that we append one value at a time.
For some reason (yet to be determined) it seems that in compaction outputs,
we often append only a single value at a time to the output writers.

After fixing this, we saw that compactions were going substantially faster in
test workload.

I also ran a micro-benchmark which stresses PLAIN-encoded ints:

  ./build/latest/cfile-test  --gtest_filter=\*100MFileNullable\*

  Before:
  I0926 16:11:13.560389  4332 cfile-test.cc:260] Time spent writing 100m nullable ints: real 6.279s     user 1.634s     sys 0.082s

  After:
  I0926 16:13:03.639183  4834 cfile-test.cc:261] Time spent writing 100m nullable ints: real 6.592s     user 1.438s     sys 0.103s

(13% improvement in CPU).

Change-Id: Id9bf4f78810b9ab1fbd708917d396d22c8bfc34a