WIP: kernel_stack_watchdog: avoid unnecessary memory fences
Per the TODO in the code, this switches from using an Acquire_Store to a simple
"StoreStore" barrier, which is free on x86. I benchmarked before and after
and the removal of the 'mfence' instruction improved performance by a significant
factor.

WIP because the storestore barrier doesn't seem to work in TSAN

Change-Id: Ie8351aebc158896571fc302b7580ea0a0fcfd5e3