KUDU-969 (part 1): simplify tablet_bootstrap-test
This cleans up some of the code in tablet_bootstrap-test to make it
easier to follow. We no longer append REPLICATE batches which contain
both an UPDATE and an INSERT in the same batch -- these turned out to
be somewhat problematic, since we couldn't actually apply the UPDATE
part without having some data in the table first.

Additionally, a lot of the numeric parameters are now more explicit --
either by use of named constants for parameters or a struct with
named parameters.

The cleaned up test mostly uses simple INSERTs. This exposes the
fact that we were lacking some actual coverage of the bootstrap code.
I plan to look at coverage analysis and add some more tests as a
follow-up part of KUDU-969 to ensure that we're covering all of the
cases in TabletBootstrap itself.

Change-Id: Id67c9c09d6fd7e62feb0e90577897de206e871e1