all_types-itest: fix placement of assertion on number of rows
Previously the assertion was in the wrong loop -- this worked in
the case that all of the rows for a tablet came back in the first
batch from a scanner. If the server was slow for some reason,
then it would often return "early" before retrieving all rows,
resulting in the appearance that some rows went missing.

Change-Id: I956fdf26773e6f57f542d5ed03e32f7817e5758d