KUDU-1094. debug-util should handle when signal handlers are already in use
The debug util code which allows stack tracing a target thread relies
on installing a signal handler. Previously, we would crash if that
signal handler was already in use.

Now, we handle the case when it's in use by disabling the functionality.
Additionally, there's a new API which we can use to allow the client
to pick a different signal handler if they are making use of the default
signal.

I also changed the default from SIGUSR1 to SIGUSR2, since SIGUSR1 is
used by the JVM.

Change-Id: Ic3035d0a442fcfeb05d78006a291240c9adccdbc