KUDU-1104. Fix a race which caused a crash in meta_cache
We were missing a lock in one place when accessing the internal map of the meta
cache.

None of the existing tests stressed multi-threaded population of the meta
cache. This patch adds a new test which creates a table with a lot of tablets,
then spawns a bunch of threads which start scanning at the same time at
different points in the table.

This new test pointed out the bug in TSAN on every run, and crashed in RELEASE
mode after 5-10 iterations.

Change-Id: I6ad5648ee386f028d45631349494a1dd0f90ce10