client: return a client-side error if trying to buffer too much data
Previously, there was no limit on how much data we could attempt to buffer in a
single batch. This would cause RPC errors on the server side, lots of retries,
etc, if you ended up buffering too much data.

Now we just return a Status::Incomplete() if we try to buffer too much between
flushes. This was discovered as part of the test I'm working on for KUDU-665,
in which I was writing very large batches to a table with 1000+ columns.

Currently, I hard-coded the buffer size at 7MB -- lower than the max RPC size
of 8MB, to account for a bit of overhead. Making it based on the RPC size flag
didn't seem to make sense, since it's the _server-side_ flag that matters, and
the client wouldn't know how that is set.

Incidentally, this also helps us get a bit closer to supporting
FlushMode::AUTO_FLUSH_BACKGROUND, but I didn't want to scope-creep
and start building that.

Change-Id: I50278e4eb070fde360e6b133cf1a1b7a333ccfac