WIP: KUDU-279 part 1. Avoid race where server shuts down keepalive when client sends a call
This WIP patch changes the RPC protocol so that, when the server decides to
disconnect an idle client, it first sends a special "GO_AWAY" message,
following the design used by HTTP 2.0. After sending this message, the server
stops accepting new calls on this connection, and waits a short amount of time
before actually disconnecting the socket.

The GO_AWAY message will also include the call ID of the last accepted call.
This allows the client to know which requests can be safely retried after
creating a new connection.

Currently, this is half-baked: the server sends the message but doesn't fill
in the call ID field. The client just disconnects upon receiving it and doesn't
do any automatic retries.

Change-Id: I81c1b7382fdbcf12ef02f58b58cb995b055c89c2