trace_event: fix JSON escaping of newlines
We were previously using the BackslashEscape utility function which places '\'
before special characters. This isn't actually what we want in order to escape
characters like newlines, etc, since we still ended up with a newline in the
text.

This broke JSON parsing if any trace events included a newline in their
parameters.

Change-Id: Ic289737026688e6491f89fd199ca31b5be02c83a