WIP: KUDU-1093: new scanner API with an encapsulated Batch object
Putting this up for early review and a test run. Adds a new API for
scanner results which encapsulates the result batch, and also
will be able to allow access to the results without constructing
an object-per-row.

Needs more fleshing out, docs on lifecycle, etc, but think overall
structure is headed in the right direction.

Change-Id: I29fd4fbb8b906ffa591853ab625ac4b089da4bc9