KUDU-1259: new scanner API with an encapsulated Batch object
This adds a new API for scanner results which encapsulates the result batch,
allowing the caller to access the rows one row at a time, rather than
constructing a vector<KuduRowResult>. This is important in the case that the
result rows are small or empty (for example an empty projection, or scanning a
single int8 column). In those cases, a single batch may return millions or even
tens of millions of rows, in which case the vector<KuduRowResult> was taking up
tens or hundreds of MBs of memory.

This refactoring ends up transferring the RpcController into the returned
RowBatch object, so it will actually be feasible to use this to avoid
copying strings in Impala -- we can simply attach the KuduScanBatch
to the Impala RowBatch to tie the lifecycle of indirect data to the
lifecycle of the rows in Impala.

I made the appropriate small change in Impala to use the new API and
verified that a SELECT COUNT(*) query which used to take 40+GB of RAM
per server now only uses a few MB. Performance also improved about 10%
for this query, likely due to less allocator pressure and page faults.

The new KuduScanBatch class fits the C++ "iterable sequence" concept,
and thus works with the C++11 range-for loop. Unfortunately it doesn't
seem to work directly with BOOST_FOREACH, though I couldn't quite
figure out the right incantation for that to work.

Change-Id: I29fd4fbb8b906ffa591853ab625ac4b089da4bc9