transaction: reduce initial arena block size
Currently transactions start with a 32KB arena. This allocation
is large enough that it will often have to go to the tcmalloc central
free list and cause contention.

In practice, many use cases perform lots of small writes with only
a few operations per batch. In these cases, 32KB is way more than
will be needed in the arena. Starting with a 1KB arena block should
be sufficient.

Change-Id: I9250ffb7afd87c274ea29e1d1bf8daf6cafb8f28