KUDU-1410 (part 3): rpc: look up the RpcMethodInfo from the reactor
This makes the reactor code look up the RpcMethodInfo associated with a call,
and moves the latency histogram collection code into the InboundCall class
instead of RpcContext. This cleans up RpcContext so now it is only responsible
for user-facing APIs and not responsible for collecting metrics.

This change may also be useful in the future if we want to start queueing calls
differently based on the particular method being invoked. We now have an easy
way for the RPC service definition to propagate up appropriate attributes
all the way to the reactor.

Change-Id: I1bed4917ca05f46c829d1733270bfba3ec20d699