WIP: KUDU-1410 (part 5): rpc: sample each RPC method in several latency buckets, expose on /rpcz
This changes RpczStore to keep several sample traces for each RPC, categorized
into several buckets (fast, slow, slower). Currently the thresholds for those
buckets are hard-coded, but follow-up work could make them adaptive based on
previous RPC percentiles, or user-configurable. Having tested this with a
YCSB workload, I found that they are useful even in this simple implementation.

WIP because this needs some new tests, though I've been testing this on a
YCSB workload and it is relatively useful.

Change-Id: I0eadb3f7035b5b156cb624ce50876876a5698b17