KUDU-1353: remove per-tablet replica locations cache
This is the simpler alternative to rebuilding the caches on metadata load.
Instead of using the caches to retrieve tserver RPC addresses, we'll look
the addresses up on the spot using the TSDescriptor cache. The two caches
should be equally coherent, as both were populated as a result of the master
receiving TS heartbeats.

I've also done away with the concept of a 'stale' locations response. The
distinction can be real, but it's confusing, and more importantly, clients
never took advantage of it. And if we're going to persist TSDescriptor data
in master state as has been recently discussed, the distinction becomes even
less useful.

Change-Id: I6376b5307f75f5d505b33a5ff4262da619cd1d8d