thread: enable stack watchdog for thread creation
In stress workloads, I see warnings fairly frequently that thread creation
is slow (on the order of 1 second). It's not clear what triggers this
behavior, yet, so this patch adds a stack watchdog to grab the kernel stack
during slow thread creation.

The only subtelty here is that starting the stack watchdog itself requires
starting a thread, so to avoid a self-deadlock, we need to start the
watchdog thread with a flag disabling watching itself.

Change-Id: Ia423084d01716e5d4884d17cdf836a43489a6ebc